requirements.txt
transformers
pillow
torch
datasets
streamlit
faiss-cpu
scikit-image
matplotlib
scikit-learn